requirements.txt
confluent-kafka==2.5.3
websocket-client==1.8.0
redis==4.9.1
